profile - password and profile picture update
- Styling to match resources - work in progress
